feat(main): implement sort, move page controller to controllers
